+++ src/columns/docker.rs
@@ -40,7 +40,7 @@ impl Docker {
         Docker {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
             containers,
@@ -79,7 +79,7 @@ impl Docker {
         Docker {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
             containers,