+++ src/columns/gid.rs
@@ -18,7 +18,7 @@ impl Gid {
         Gid {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
         }