+++ src/columns/start_time.rs
@@ -21,7 +21,7 @@ impl StartTime {
         StartTime {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
         }