+++ src/columns/tcp_port.rs
@@ -26,7 +26,7 @@ impl TcpPort {
         TcpPort {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
             #[cfg(target_os = "linux")]