+++ src/columns/usage_mem.rs
@@ -20,7 +20,7 @@ impl UsageMem {
         UsageMem {
             fmt_contents: HashMap::new(),
             raw_contents: HashMap::new(),
-            max_width: cmp::max(header.len(), unit.len()),
+            max_width: 0,
             header,
             unit,
             mem_total: get_mem_total(),