+++ src/util.rs
@@ -1,5 +1,6 @@
 use crate::column::Column;
 use crate::config::{ConfigColumnAlign, ConfigSearchLogic};
+use unicode_width::UnicodeWidthStr;
 
 pub enum KeywordClass {
     Numeric,
@@ -76,10 +77,10 @@ pub fn classify(keyword: &str) -> KeywordClass {
 
 pub fn expand(x: &str, len: usize, align: &ConfigColumnAlign) -> String {
     match align {
-        ConfigColumnAlign::Left => format!("{}{}", x, " ".repeat(len - x.chars().count())),
-        ConfigColumnAlign::Right => format!("{}{}", " ".repeat(len - x.chars().count()), x),
+        ConfigColumnAlign::Left => format!("{}{}", x, " ".repeat(len - UnicodeWidthStr::width(x))),
+        ConfigColumnAlign::Right => format!("{}{}", " ".repeat(len - UnicodeWidthStr::width(x)), x),
         ConfigColumnAlign::Center => {
-            let space = len - x.chars().count();
+            let space = len - UnicodeWidthStr::width(x);
             let left = space / 2;
             let right = space / 2 + space % 2;
             format!("{}{}{}", " ".repeat(left), x, " ".repeat(right))