+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Added] column description to `--list` output
 * [Changed] sort indicator refine
+* [Fixed] wrong column width calculation about full-width charactors
 
 ## [v0.5.7](https://github.com/dalance/procs/compare/v0.5.6...v0.5.7) - 2019-03-05